refactored code into multiple scripts
