+++ README.md
@@ -1,2 +1,14 @@
 # intelligentroutes
 COS30018 IS vehicle routing problem
+ To run the code you'll need the following :
+ - distance.py : script to multiple functions to calculate distances
+ - vehicle.py : script for DA with the respecting constraints
+ - routing.py : script for initial searching algorithm (greedy) script
+ - optimizer.py : MRA or main optimization algorithm script
+ - visualization.py : script with multiple function to visualize the results
+ - data_parser.py : script forparse uploaded json file and catch any error 
+ - generator.py : script togenerate random coordinates, demands, time_window and save it to coordinates.json
+ - main.py : script to display gui via streamlit
+ - coordinates.json : randomly generated data
+ - vehicle_data.json : custom user input data
+ - set_coordinates.json : set generated data
