+++ coordinates.json
@@ -5,248 +5,28 @@
             145.0669
         ],
         [
-            -37.83636646246559,
-            145.06954307544345
+            -37.84496479436458,
+            145.10934407330788
         ],
         [
-            -37.82588256208366,
-            145.05594886540038
+            -37.82222011342774,
+            145.04877915572462
         ],
         [
-            -37.80352983099736,
-            145.10289535429385
+            -37.79271942733241,
+            145.0766091485864
         ],
         [
-            -37.809574553900184,
-            145.07554766850836
+            -37.840409483932945,
+            145.1064510676896
         ],
         [
-            -37.84434366064344,
-            145.09462200738545
-        ],
-        [
-            -37.81081617995398,
-            145.04885810828642
-        ],
-        [
-            -37.8303919393057,
-            145.11002057551647
-        ],
-        [
-            -37.794559850688756,
-            145.05566407880832
-        ],
-        [
-            -37.82343328989943,
-            145.05426121469202
-        ],
-        [
-            -37.806687479889234,
-            145.1071060448131
-        ],
-        [
-            -37.80047913021299,
-            145.07677559556322
-        ],
-        [
-            -37.80999065413811,
-            145.07845929894495
-        ],
-        [
-            -37.80641173727996,
-            145.08346701664246
-        ],
-        [
-            -37.83601508073989,
-            145.08878190759557
-        ],
-        [
-            -37.819487127076314,
-            145.08047677624901
-        ],
-        [
-            -37.81858020017025,
-            145.08395126411045
-        ],
-        [
-            -37.8074929392582,
-            145.07173688304903
-        ],
-        [
-            -37.815830230034045,
-            145.03244176278142
-        ],
-        [
-            -37.79429275873893,
-            145.06647575297563
-        ],
-        [
-            -37.841175016318495,
-            145.02534734712384
-        ],
-        [
-            -37.81477087950779,
-            145.06636774211503
-        ],
-        [
-            -37.84234293596256,
-            145.04510680249078
-        ],
-        [
-            -37.84545110056714,
-            145.07159502973172
-        ],
-        [
-            -37.83867663434986,
-            145.02306516598628
-        ],
-        [
-            -37.83399511675743,
-            145.10255200425732
-        ],
-        [
-            -37.79442735162902,
-            145.0366404558845
-        ],
-        [
-            -37.844795567866676,
-            145.0687293417342
-        ],
-        [
-            -37.82905231938748,
-            145.09634806949558
-        ],
-        [
-            -37.812086594042256,
-            145.0954030572462
-        ],
-        [
-            -37.82753770101012,
-            145.0267206448408
-        ],
-        [
-            -37.84124366961064,
-            145.04426896019143
-        ],
-        [
-            -37.80458614882061,
-            145.10368376055715
-        ],
-        [
-            -37.84366401371481,
-            145.111708535446
-        ],
-        [
-            -37.81337342094792,
-            145.05657781272566
-        ],
-        [
-            -37.79902016427806,
-            145.04908170060338
-        ],
-        [
-            -37.84687182430611,
-            145.07108845650828
-        ],
-        [
-            -37.82142860496712,
-            145.03113211606066
-        ],
-        [
-            -37.82777638090036,
-            145.09587772453364
-        ],
-        [
-            -37.84683779123113,
-            145.09957389667466
-        ],
-        [
-            -37.82937388600226,
-            145.06222726862703
-        ],
-        [
-            -37.82574068948524,
-            145.0821584402469
-        ],
-        [
-            -37.832614807763186,
-            145.0885248474634
-        ],
-        [
-            -37.811277096531846,
-            145.0498413469919
-        ],
-        [
-            -37.832539510304905,
-            145.06424932712855
-        ],
-        [
-            -37.81607591493016,
-            145.08897435036175
-        ],
-        [
-            -37.84497331320966,
-            145.07792141706395
-        ],
-        [
-            -37.81963805960871,
-            145.05971917980332
-        ],
-        [
-            -37.79603774678851,
-            145.058478069467
-        ],
-        [
-            -37.817869689482976,
-            145.0238561205345
+            -37.79821500765467,
+            145.0760911506649
         ]
     ],
     "demands": [
         0,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
-        1,
         1,
         1,
         1,
@@ -259,104 +39,8 @@
             999
         ],
         [
-            17,
-            22
-        ],
-        [
             12,
             17
-        ],
-        [
-            22,
-            5
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            22,
-            5
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            5,
-            12
-        ],
-        [
-            22,
-            5
-        ],
-        [
-            22,
-            5
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            22,
-            5
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            5,
-            12
-        ],
-        [
-            5,
-            12
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            22,
-            5
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            5,
-            12
         ],
         [
             12,
@@ -369,86 +53,6 @@
         [
             17,
             22
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            5,
-            12
-        ],
-        [
-            5,
-            12
-        ],
-        [
-            5,
-            12
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            22,
-            5
-        ],
-        [
-            5,
-            12
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            5,
-            12
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            17,
-            22
-        ],
-        [
-            12,
-            17
-        ],
-        [
-            22,
-            5
-        ],
-        [
-            5,
-            12
         ],
         [
             12,
